+++ bin/api_diff.ml
@@ -1,45 +1,83 @@
 let tool_name = "api-diff"
 
-let run (`Main_module main_module) (`Ref_cmi reference) (`Current_cmi current) =
+type mode = Unwrapped | Wrapped of string | Cmi
+
+let both_directories reference current =
+  match (Sys.is_directory reference, Sys.is_directory current) with
+  | true, true -> Ok true
+  | false, false -> Ok false
+  | _ ->
+      Error
+        "Arguments must either both be directories or both single .cmi files."
+
+let print_warning main_module unwrapped =
+  match (main_module, unwrapped) with
+  | None, false -> ()
+  | Some _, false ->
+      Printf.eprintf
+        "%s: --main-module is ignored when diffing single .cmi files\n"
+        tool_name
+  | None, true ->
+      Printf.eprintf
+        "%s: --unwrapped is ignored when diffing single .cmi files\n" tool_name
+  | Some _, true ->
+      Printf.eprintf
+        "%s: --main-module and --unwrapped are ignored when diffing single \
+         .cmi files\n"
+        tool_name
+
+let mode ~reference ~current ~main_module ~unwrapped =
   let open CCResult.Infix in
-  let* reference_sig, current_sig, module_name =
-    match
-      (Sys.is_directory reference, Sys.is_directory current, main_module)
-    with
-    | true, true, Some main_module ->
+  let* both_dirs = both_directories reference current in
+  match (both_dirs, main_module, unwrapped) with
+  | true, Some main_module, false -> Ok (Wrapped main_module)
+  | true, None, true -> Ok Unwrapped
+  | false, main_module, unwrapped ->
+      print_warning main_module unwrapped;
+      Ok Cmi
+  | true, _, _ ->
+      Error
+        "Either --main-module or --unwrapped must be provided when diffing \
+         entire libraries."
+
+let run (`Main_module main_module) (`Unwrapped_library unwrapped)
+    (`Ref_cmi reference) (`Current_cmi current) =
+  let open CCResult.Infix in
+  let* reference_map, current_map =
+    let* curr_mode = mode ~reference ~current ~main_module ~unwrapped in
+    match curr_mode with
+    | Wrapped main_module ->
         let main_module = String.capitalize_ascii main_module in
-        let+ reference_sig = Api_watch.Library.load ~main_module reference
-        and+ current_sig = Api_watch.Library.load ~main_module current in
-        let module_name = String.capitalize_ascii main_module in
-        (reference_sig, current_sig, module_name)
-    | false, false, main_module ->
-        let () =
-          match main_module with
-          | None -> ()
-          | Some _ ->
-              Printf.eprintf
-                "%s: --main-module ignored when diffing single .cmi files\n"
-                tool_name
-        in
+        let+ reference_map = Api_watch.Library.load ~main_module reference
+        and+ current_map = Api_watch.Library.load ~main_module current in
+        (reference_map, current_map)
+    | Unwrapped ->
+        let+ reference_map = Api_watch.Library.load_unwrapped reference
+        and+ current_map = Api_watch.Library.load_unwrapped current in
+        (reference_map, current_map)
+    | Cmi ->
         let+ reference_cmi, _ = Api_watch.Library.load_cmi reference
         and+ current_cmi, module_name = Api_watch.Library.load_cmi current in
-        (reference_cmi, current_cmi, module_name)
-    | true, false, _ | false, true, _ ->
-        Error
-          "Arguments must either both be directories or both single .cmi files."
-    | true, true, None ->
-        Error "--main-module must be provided when diffing entire libraries."
+        let reference_map =
+          Api_watch.String_map.singleton module_name reference_cmi
+        in
+        let current_map =
+          Api_watch.String_map.singleton module_name current_cmi
+        in
+        (reference_map, current_map)
   in
-  let diff =
-    Api_watch.Diff.interface ~module_name ~reference:reference_sig
-      ~current:current_sig
+  let diff_map =
+    Api_watch.Diff.library ~reference:reference_map ~current:current_map
+    |> Api_watch.String_map.bindings
+    |> List.filter_map (fun (_, v) -> v)
   in
-  match diff with
-  | None -> Ok 0
-  | Some diff ->
+  let has_changes = not (List.is_empty diff_map) in
+  List.iter
+    (fun diff ->
       let text_diff = Api_watch.Text_diff.from_diff diff in
-      Api_watch.Text_diff.With_colors.pp Format.std_formatter text_diff;
-      Ok 1
+      Api_watch.Text_diff.With_colors.pp Format.std_formatter text_diff)
+    diff_map;
+  if has_changes then Ok 1 else Ok 0
 
 let named f = Cmdliner.Term.(app (const f))
 
@@ -53,6 +91,15 @@ let main_module =
     (fun x -> `Main_module x)
     Cmdliner.Arg.(
       value & opt (some string) None & info ~doc ~docv [ "main-module" ])
+
+let unwrapped_library =
+  let doc =
+    "Loads a library without a main module. Ignored when diffing single \
+     $(b,.cmi) files."
+  in
+  named
+    (fun x -> `Unwrapped_library x)
+    Cmdliner.Arg.(value & flag & info ~doc [ "unwrapped" ])
 
 let ref_cmi =
   let docv = "REF_CMI_FILES" in
@@ -79,7 +126,9 @@ let info =
   Cmd.info tool_name ~version:"%%VERSION%%" ~exits:Cmd.Exit.defaults
     ~doc:"List API changes between two versions of a library"
 
-let term = Cmdliner.Term.(const run $ main_module $ ref_cmi $ current_cmi)
+let term =
+  Cmdliner.Term.(
+    const run $ main_module $ unwrapped_library $ ref_cmi $ current_cmi)
 
 let () =
   Fmt_tty.setup_std_outputs ();